first pass at adding new damage and weapon stats
currently this will print the damage chart (who damaged whom, for how much) automatically with stats, BUT the weapon charts (damage that a specific player dealt/received from each weapon) only print automatically for that specific player if they were ingame. this is because with two tables for every player, a 4v4 match would result in 16 tables, which would be insane console spam. every table does, however, go into the buffer that is printed by the discord webhook.

a necessary enhancement will be to allow viewing these tables AFTER a match is complete (i.e. type the stats command and it shows stats from the previous match), meaning they will need to be saved somewhere. but what if people disconnect, client numbers change, etc.? might be helpful to use this opportunity to track stats somewhere other than client structs so that they aren't lost if someone ragequits or disconnects.

a further enhancement might be that, if all this data is going to be saved outside of the client struct anyway, why not go one step further and save it into the db? e.g. so you could track average flag hold time, etc. across many matches

also, elo bot will probably need to be updated to support all this